Update facebook ad id Update gradle refactor app
