+++ app/build.gradle
@@ -1,19 +1,12 @@
-buildscript {
-    repositories {
-        maven { url 'https://maven.fabric.io/public' }
-    }
-
-    dependencies {
-        classpath 'io.fabric.tools:gradle:1.24.4'
-    }
-}
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
 apply plugin: 'com.google.firebase.firebase-perf'
 apply plugin: 'io.fabric'
 
 repositories {
     maven { url 'https://maven.fabric.io/public' }
     google()
+    mavenCentral()
 }
 
 
@@ -27,12 +20,12 @@ android {
             v2SigningEnabled false
         }
     }
-    compileSdkVersion 27
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
     defaultConfig {
         applicationId "com.deeppandya.appmanager"
         minSdkVersion 15
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 13
         versionName "1.0.13"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -56,63 +49,82 @@ android {
     }
 }
 
-ext {
-    supportLibraryVersion = '27.0.0'
-    playServicesLibraryVersion = '11.4.2'
-}
-
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
 
+    configurations {
+        all*.exclude group: 'com.android.support', module: 'support-v13'
+    }
+
     //android support
-    compile "com.android.support:appcompat-v7:${supportLibraryVersion}"
-    compile "com.android.support:support-v4:${supportLibraryVersion}"
-    compile "com.android.support:design:${supportLibraryVersion}"
-    compile "com.android.support:recyclerview-v7:${supportLibraryVersion}"
-    compile "com.android.support:cardview-v7:${supportLibraryVersion}"
+    implementation "com.android.support:appcompat-v7:28.0.0"
+    implementation "com.android.support:support-v4:28.0.0"
+    implementation "com.android.support:design:28.0.0"
+    implementation "com.android.support:recyclerview-v7:28.0.0"
+    implementation "com.android.support:cardview-v7:28.0.0"
+    implementation "com.android.support:customtabs:28.0.0"
 
     //Firebase
-    compile "com.google.firebase:firebase-core:${playServicesLibraryVersion}"
-    compile "com.google.firebase:firebase-crash:${playServicesLibraryVersion}"
-    compile "com.google.firebase:firebase-messaging:${playServicesLibraryVersion}"
-    compile "com.google.firebase:firebase-config:${playServicesLibraryVersion}"
-    compile "com.google.firebase:firebase-invites:${playServicesLibraryVersion}"
-    compile "com.google.firebase:firebase-database:${playServicesLibraryVersion}"
-    compile "com.google.firebase:firebase-perf:${playServicesLibraryVersion}"
-    compile "com.google.firebase:firebase-auth:${playServicesLibraryVersion}"
+    implementation "com.google.firebase:firebase-core:16.0.8"
+    implementation "com.google.firebase:firebase-crash:16.2.1"
+    implementation "com.google.firebase:firebase-messaging:17.5.0"
+    implementation "com.google.firebase:firebase-config:16.4.1"
+    implementation "com.google.firebase:firebase-invites:16.1.1"
+    implementation "com.google.firebase:firebase-database:16.1.0"
+    implementation "com.google.firebase:firebase-perf:16.2.4"
+    implementation "com.google.firebase:firebase-auth:16.2.0"
 
     //google play services
-    compile "com.google.android.gms:play-services-ads:${playServicesLibraryVersion}"
+//    implementation "com.google.android.gms:play-services-ads:17.2.0"
 
     //External dependency
-    compile('com.crashlytics.sdk.android:crashlytics:2.6.8@aar') {
+    implementation('com.crashlytics.sdk.android:crashlytics:2.9.9@aar') {
         transitive = true
     }
 
-    compile 'com.afollestad.material-dialogs:core:0.9.2.3'
-    compile 'com.github.bumptech.glide:glide:3.8.0'
-    compile 'com.android.support:multidex:1.0.2'
-    compile 'com.evernote:android-job:1.0.7'
-    compile 'com.appbrain:appbrain-sdk:13.22@aar'
-    compile 'com.android.support.constraint:constraint-layout:1.1.0-beta3'
-    compile 'org.jokar:permissiondispatcher:2.0.8'
+    implementation 'com.afollestad.material-dialogs:core:0.9.2.3'
+    implementation 'com.github.bumptech.glide:glide:4.9.0'
+    implementation 'com.android.support:multidex:1.0.3'
+    implementation 'com.evernote:android-job:1.0.7'
+    implementation 'com.appbrain:appbrain-sdk:13.22@aar'
+    implementation 'com.android.support.constraint:constraint-layout:2.0.0-alpha3'
+    implementation 'org.jokar:permissiondispatcher:2.0.8'
 
-    compile 'com.facebook.android:facebook-android-sdk:4.27.0'
-    compile 'com.facebook.android:audience-network-sdk:4.26.1'
+    implementation 'com.facebook.android:facebook-android-sdk:4.27.0'
+    implementation 'com.facebook.android:audience-network-sdk:4.26.1'
 
-    compile 'com.codemybrainsout.rating:ratingdialog:1.0.7'
-    compile 'com.github.aakira:expandable-layout:1.6.0@aar'
+    implementation 'com.codemybrainsout.rating:ratingdialog:1.0.7'
+    implementation 'com.github.aakira:expandable-layout:1.6.0@aar'
 
-    compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
-    compile 'io.reactivex.rxjava2:rxjava:2.1.5'
+    implementation 'com.google.code.gson:gson:2.8.5'
 
-    debugCompile 'com.squareup.leakcanary:leakcanary-android:1.5.4'
-    releaseCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.4'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.0.1'
+    implementation 'io.reactivex.rxjava2:rxjava:2.1.5'
 
-    testCompile 'junit:junit:4.12'
+    debugImplementation 'com.squareup.leakcanary:leakcanary-android:1.5.4'
+    releaseImplementation 'com.squareup.leakcanary:leakcanary-android-no-op:1.5.4'
+
+    // Room components
+    implementation 'android.arch.persistence.room:runtime:1.1.1'
+    annotationProcessor 'android.arch.persistence.room:compiler:1.1.1'
+
+    // Lifecycle components
+    implementation 'android.arch.lifecycle:extensions:1.1.1'
+    annotationProcessor 'android.arch.lifecycle:compiler:1.1.1'
+
+    // Livedata
+    implementation 'android.arch.lifecycle:livedata-core:1.1.1'
+
+    testImplementation 'junit:junit:4.12'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    implementation 'com.google.dagger:dagger:2.11'
+    annotationProcessor 'com.google.dagger:dagger-compiler:2.11'
 }
 
 apply plugin: 'com.google.gms.google-services'
+
+apply plugin: 'kotlin-android-extensions'
