--- app/src/main/java/com/deeppandya/appmanager/util/BulletTextUtil.java
@@ -1,39 +1,0 @@
-package com.deeppandya.appmanager.util;
-
-import android.content.Context;
-import android.text.Html;
-import android.text.Spannable;
-import android.text.SpannableString;
-import android.text.SpannableStringBuilder;
-import android.text.Spanned;
-import android.text.style.BulletSpan;
-
-/**
- * Created by d_pandya on 3/10/17.
- */
-
-public class BulletTextUtil {
-
-    private static final String SPACE = " ";
-    private static final String BULLET_SYMBOL = "&#8226";
-    private static final String EOL = System.getProperty("line.separator");
-    private static final String TAB = "\t";
-
-    public BulletTextUtil() {
-    }
-
-    public static String getBulletList(String header, String[] items) {
-        StringBuilder listBuilder = new StringBuilder();
-        if (header != null && !header.isEmpty()) {
-            listBuilder.append(header + EOL + EOL);
-        }
-        if (items != null && items.length != 0) {
-            for (String item : items) {
-                Spanned formattedItem = Html.fromHtml(BULLET_SYMBOL + SPACE + item);
-                listBuilder.append(TAB + formattedItem + EOL);
-            }
-        }
-        return listBuilder.toString();
-    }
-
-}
